stage-2-production/services/notification-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiokafka==0.12.0
pydantic==2.10.3
prometheus-client==0.21.1
python-dotenv==1.0.1
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
twilio==9.0.0
